test/pytest/requirements.txt
psycopg2-binary
orderedset
sortedcontainers
pytest>=3.9.1
git+https://github.com/bitaps-com/pybgl
requests




